Retours sur le premier TP

+++ quiz-api/app.py
@@ -1,6 +1,5 @@
-import json
 from flask import Flask, request
-from importlib_metadata import method_cache
+from flask_cors import CORS
 from utils.jwt_utils import build_token
 from utils.jwt_utils import decode_token
 import service.questionService as questionService
@@ -8,8 +7,8 @@ import service.participationService as participationService
 import service.quizInfoService as quizInfoService
 from model.participation import serialize
 
-
 app = Flask(__name__)
+CORS(app)
 
 @app.route('/')
 def hello_world():
@@ -18,7 +17,6 @@ def hello_world():
 
 @app.route('/quiz-info', methods=['GET'])
 def GetQuizInfo():
-	#return {"size": 0, "scores": []}, 200
 	return quizInfoService.getQuizInfo()
 
 @app.route('/login', methods=['POST'])
@@ -48,8 +46,12 @@ def GetQuestion(question_position):
 
 @app.route('/questions/<question_position>', methods=['PUT'])
 def UpdateQuestion(question_position):
-	newQuestion = request.get_json()
-	return questionService.updateQuestionByPosition(question_position,newQuestion)
+	try:
+		if (decode_token((request.headers.get('Authorization'))[7:])=='quiz-app-admin'):
+			newQuestion = request.get_json()
+			return questionService.updateQuestionByPosition(question_position,newQuestion)
+	except Exception:
+		return '',400
 
 @app.route('/questions/<question_position>', methods=['DELETE'])
 def DeleteQuestion(question_position):
@@ -73,7 +75,11 @@ def SaveParticipation():
 
 @app.route('/participations', methods=['DELETE'])
 def DeleteParticipation():
-	return quizInfoService.deleteAllQuizInfo()
-		
+	try:
+		if (decode_token((request.headers.get('Authorization'))[7:])=='quiz-app-admin'):
+			return quizInfoService.deleteAllQuizInfo()
+	except Exception:
+		return '',400
+
 if __name__ == "__main__":
-	app.run(ssl_context='adhoc')
+	app.run()
